Refactor: Split utility functions
Split utility functions to allow for easier reading of the methods in each file.
This will better prepare the files for the addition of the favorite functionality
the database was implemented for.
