+++ app/src/main/java/com/omegaspocktari/movieprojectone/utilities/NetworkUtils.java
@@ -1,0 +1,91 @@
+package com.omegaspocktari.movieprojectone.utilities;
+
+import android.util.Log;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.nio.charset.Charset;
+
+/**
+ * Created by ${Michael} on 4/18/2017.
+ */
+
+public class NetworkUtils {
+
+    private static final String LOG_TAG = NetworkUtils.class.getSimpleName();
+
+    protected static URL createUrl(String stringUrl) {
+        URL url = null;
+
+        try {
+            url = new URL(stringUrl);
+        } catch (MalformedURLException e) {
+            Log.e(LOG_TAG, "Problem constructing URL Object [TMDbJsonUtils].", e);
+        }
+
+        return url;
+    }
+
+    protected static String makeHTTPRequest(URL url) throws IOException {
+        String jsonResponse = "";
+
+        // End HTTP Request if the url is null
+        if (url == null) {
+            return jsonResponse;
+        }
+
+        // URL Connection to acquire data from the web and the variable we'll be storing it in
+        HttpURLConnection urlConnection = null;
+        InputStream inputStream = null;
+
+        try {
+            urlConnection = (HttpURLConnection) url.openConnection();
+            urlConnection.setReadTimeout(1000);
+            urlConnection.setConnectTimeout(15000);
+            urlConnection.setRequestMethod("GET");
+            urlConnection.connect();
+
+            if (urlConnection.getResponseCode() == 200) {
+                inputStream = urlConnection.getInputStream();
+                jsonResponse = readFromStream(inputStream);
+            }
+        } catch (IOException e) {
+            Log.e(LOG_TAG, "Error making an HTTP Request... ", e);
+            e.printStackTrace();
+        } finally {
+            if (urlConnection != null) {
+                urlConnection.disconnect();
+            } if (inputStream != null) {
+                inputStream.close();
+            }
+        }
+
+        return jsonResponse;
+    }
+
+    private static String readFromStream(InputStream inputStream) throws IOException {
+        StringBuilder output = new StringBuilder();
+
+        if (inputStream != null) {
+            // Byte stream to character stream. Decodes bytes utilizing the given charset.
+            InputStreamReader inputStreamReader = new InputStreamReader(inputStream, Charset.forName("UTF-8"));
+
+            //
+            // Good to wrap buffered reader around Readers whose read() operations may be costly.
+            BufferedReader reader = new BufferedReader(inputStreamReader);
+            // Simple loop to read through the inputStream.
+            String line = reader.readLine();
+            while (line != null) {
+                output.append(line);
+                line = reader.readLine();
+            }
+        }
+
+        return output.toString();
+    }
+}
